requirements.txt
gradio>=3.0
openai>=1.0.0
python-dotenv>=0.21.0
pygments>=2.14.0
